refactor: update package management handlers and adjust server variable imports
