+++ apps/electron/src/main/modules/agent/package/install-package-manager.ts
