+++ apps/electron/src/main/modules/agent/package/package-version-resolver.ts
