+++ apps/electron/src/main/modules/agent/shared/server-variable-utils.ts
