+++ apps/electron/src/renderer/components/agent/create/ChatInterface.tsx
@@ -20,7 +20,7 @@ import {
   TooltipProvider,
 } from "@mcp_router/ui";
 import { AgentConfig, MCPServerConfig } from "@mcp_router/shared";
-import { extractServerVariables } from "@/main/modules/mcp-core/server/shared/server-variable-utils";
+import { extractServerVariables } from "@/main/modules/agent/shared/server-variable-utils";
 import {
   Card,
   CardContent,
